requirements.txt
akshare>=1.12.0
pandas>=2.0.0
PyYAML>=6.0
PyMySQL>=1.1.0
cryptography>=41.0.0
flask>=3.0.0
matplotlib>=3.7.0
scipy>=1.10.0
statsmodels>=0.14.0
PyWavelets>=1.4.0
torch>=2.0.0
scikit-learn>=1.3.0
shap>=0.43.0
xgboost>=2.0.0
lightgbm>=4.0.0
requests
python-dotenv
APScheduler
SQLAlchemy>=2.0.0
redis>=5.0.0
